[Refactor] Move clone task to data workflows
