+++ be/src/agent/agent_task.cpp
@@ -43,7 +43,7 @@
 #include "storage/snapshot_manager.h"
 #include "storage/storage_env.h"
 #include "storage/tablet_manager.h"
-#include "storage/task/engine_clone_task.h"
+#include "data_workflows/clone/engine_clone_task.h"
 #include "storage/txn_manager.h"
 #include "storage/update_manager.h"
