+++ be/src/data_workflows/clone/engine_clone_task.cpp
@@ -32,7 +32,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "storage/task/engine_clone_task.h"
+#include "data_workflows/clone/engine_clone_task.h"
 
 #include <fmt/format.h>
 #include <sys/stat.h>
